pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-saas"
version = "0.1.0"
description = "A comprehensive Django SaaS modular monorepo with 89 standalone apps"
requires-python = ">=3.12"
dependencies = [
    # Core Django
    "Django>=5.1,<6.0",
    "psycopg[binary]>=3.2",
    "gunicorn>=23.0",
    "uvicorn[standard]>=0.34",

    # API Frameworks
    "djangorestframework>=3.15",
    "django-ninja>=1.3",
    "drf-spectacular>=0.28",
    "djangorestframework-simplejwt>=5.3",

    # Authentication & Authorization
    "django-allauth[mfa,socialaccount]>=65.0",
    "dj-rest-auth>=7.0",
    "django-otp>=1.5",
    "webauthn>=2.2",
    "django-guardian>=2.4",
    "rules>=3.5",
    "djangorestframework-api-key>=3.0",
    "djangosaml2>=1.10",
    "mozilla-django-oidc>=4.0",

    # Multi-tenancy
    "django-tenants>=3.7",

    # Billing
    "dj-stripe>=2.9",

    # Email
    "django-anymail[amazon_ses,sendgrid,mailgun,postmark]>=12.0",
    "django-post-office>=3.9",
    "premailer>=3.10",

    # Notifications
    "django-notifications-hq>=1.8",
    "django-push-notifications>=3.1",

    # Background Tasks
    "celery[redis]>=5.4",
    "django-celery-beat>=2.7",
    "django-celery-results>=2.5",

    # Caching
    "django-redis>=5.4",

    # Storage
    "django-storages[boto3,google,azure]>=1.14",
    "Pillow>=11.0",
    "django-imagekit>=5.0",

    # Search
    "meilisearch>=0.33",

    # Security
    "django-cors-headers>=4.6",
    "django-csp>=4.0",
    "django-axes>=7.0",
    "django-encrypted-model-fields>=0.6",

    # Audit & History
    "django-auditlog>=3.0",
    "django-simple-history>=3.8",

    # Feature Flags
    "django-waffle>=4.2",

    # Health & Monitoring
    "django-health-check>=3.18",
    "sentry-sdk[django]>=2.19",
    "django-prometheus>=2.3",
    "structlog>=24.4",
    "opentelemetry-instrumentation-django>=0.50",

    # Admin
    "django-unfold>=0.42",
    "django-hijack>=3.6",

    # User Management
    "django-invitations>=2.1",
    "django-user-sessions>=2.0",

    # Content & CMS
    "django-markdownx>=4.0",
    "django-taggit>=6.1",

    # SEO
    "django-meta>=2.4",

    # i18n / l10n
    "django-modeltranslation>=0.19",
    "django-countries>=7.6",
    "Babel>=2.16",

    # Config
    "setuptools>=75.0,<81",
    "django-environ>=0.12",
    "whitenoise>=6.8",

    # WebSockets
    "channels[daphne]>=4.2",
    "channels-redis>=4.2",

    # OAuth Provider
    "django-oauth-toolkit>=3.0",

    # Exports
    "openpyxl>=3.1",
    "WeasyPrint>=63.0",

    # Activity
    "django-activity-stream>=2.0",

    # Rate Limiting
    "django-ratelimit>=4.1",

    # Data Privacy
    "django-cookie-consent>=0.6",

    # Database Utilities
    "django-tree-queries>=0.19",

    # Testing (also runtime for test_fixtures app)
    "factory-boy>=3.3",
    "Faker>=33.0",

    # AI/ML
    "litellm>=1.55",
    "anthropic>=0.40",
    "openai>=1.58",
    "tiktoken>=0.8",
    "pgvector>=0.3",

    # SMS
    "twilio>=9.4",

    # Webhooks
    "svix>=1.47",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
    "django-debug-toolbar>=4.4",
    "ipython>=8.30",
    "pre-commit>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["apps", "config"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "DJ"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"
